Move resources to CustomerSession/EmployeeSession folders (match URI first segment)
Rector's ApiResourceUriTemplateRector derives the first URI segment from the
namespace folder (pluralized). Putting the resources under 'Security/' forced
'/securities/' as first segment; splitting into 'CustomerSession/' and
'EmployeeSession/' matches the '/customer-sessions/' and '/employee-sessions/'
URIs cleanly.
